Set global variable _STDOUT_TYPE.
This value holds a string that describes a type of stdout. This value
changes, depending where stdout is redirected.

This feature addresses a problem a script might have if it uses the
ASCII sequences to change a foreground/background color. A script can
determine if it's writting to a terminal window or a file,
enabling/disabling colors accordingly.